# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Documentation for Ecommerce-Application",
        "title": "Ecommerce",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:3000",
    "paths": {
...
        "/vendor-edit-details": {
            "post": {
                "description": "This Handler enables Vendor to edit his/her details",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Vendor"
                ],
                "summary": "Vendor Edit Profile Details",
                "parameters": [
                    {
                        "description": "company details",
                        "name": "edit-profile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "{  \"companyName\": \"Ambani seth\",\"description\" : \"Abra ka dabra jbewijbwr\",\"logo\" : \"/home/chicmic/Downloads/test.jpg\",\"bannerImage\" : \"/home/chicmic/Downloads/test.jpg\"}"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/vendor-login": {
            "post": {
                "description": "Login Vendor with initial details in DB",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Vendor"
                ],
                "summary": "Vendor Login Handler",
                "parameters": [
                    {
                        "description": "vendor details",
                        "name": "vendor-register",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "{  \"gstNumber\" : \"29ABCDE1234F1Z5\",\"companyContact\" : \"string\"}"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/vendor-logout": {
            "delete": {
                "description": "This Handler will Log out the user",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Vendor"
                ],
                "summary": "Vendor Logout",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/vendor-product-add": {
            "post": {
                "description": "This Handler adds multiple products from vendor side",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Vendor"
                ],
                "summary": "Product Add from vendor side",
                "parameters": [
                    {
                        "description": "product description",
                        "name": "product-add",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "{   \"productName\" : \"Redmi Note 10 pro 4G\",\"productDescription\" : \"4G smart phone\",\"productPrice\" : 9999.9,\"productBrand\" : \"Redmi\",\"productCategory\" : \"mobile\",\"productImageUrl\" : \"http://dummyimage.com/169x100.png/cc0000/ffffff\"}"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/vendor-product-delete": {
            "post": {
                "description": "This Handler will delete product from the inventory",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Vendor"
                ],
                "summary": "Product Delete from Inventory",
                "parameters": [
                    {
                        "description": "product id",
                        "name": "inventory-delete",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "{  \"productId\" : \"string\"}"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/vendor-product-update": {
            "post": {
                "description": "This Handler Update Product Inventory Details",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Vendor"
                ],
                "summary": "Product Inventory Update",
                "parameters": [
                    {
                        "description": "product id and product inventory",
                        "name": "product-inventory-update",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "{  \"productId\" : \"string\",\"productInventory\" : \"float64\"}"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/vendor-register": {
            "post": {
                "description": "Registering Vendor with initial details in DB",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Vendor"
                ],
                "summary": "Vendor Register Handler",
                "parameters": [
                    {
                        "description": "vendor details",
                        "name": "vendor-register",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "{  \"gstNumber\" : \"29ABCDE1234F1Z5\",\"companyName\" : \"Sports Tak\",\"companyContact\" : \"9877370350\",\"street\" : \"saytan gali kholi number 420\",\"city\" : \"mohali\",\"state\" : \"punjab\",\"postalCode\" : \"152001\",\"country\" : \"india\"}"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/vendor-verify-otp": {
            "post": {
                "description": "Verify the OTP against the provided phone number",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Vendor"
                ],
                "summary": "Vendor Verify OTP Handler",
                "parameters": [
                    {
                        "description": "phone number and otp of the user",
                        "name": "verify-otp",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "{ \"contactNumber\" : \"string\" , \"otp\" : \"string\" }"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    }
}